Initial (and only) commit
This was intended to be an experiment in implementing a
createElement function for React that takes a ES Module.
Work on this has been stopped and this code probably doesn't
work. Now that React's hooks have been announced, I don't really see
a point in continuing to work on this, especially since this
approach has some downsides that I don't like (and hooks are
much cooler).

Inspired by
https://github.com/reactjs/react-future/blob/master/07%20-%20Returning%20State/02%20-%20Module%20Pattern.js
and
https://github.com/reactjs/react-future/blob/master/07%20-%20Returning%20State/01%20-%20Stateful%20Functions.js
initially, and later
https://github.com/reasonml/reason-react.

The implementation is simple; if the function is passed a module,
create a class from it and pass that to React's createElement,
otherwise just pass the arguments to React's createElement.
Caching the created classes is (afaik) necessary so that we
provide the same class to React's createElement each time
the same module is passed in (if it's not necessary, it should
help performance anyways).

There are a few downsides to this approach. The first is that
I do not particularly care for the implicit binding of 'this'
here. It's not obvious what 'this' is when you're in one of
the lifecycle functions. Another downside is that quite a few
of these functions have many arguments. componentDidUpdate for
example has 4, along with the implicit 'this' value. I began
converting this to follow reason-react's API, however before
I could finish, hooks were announced and those seem like a much
better abstraction (yay algebraic effects).